+++ contracts/utils/MarketsLib.sol
@@ -39,8 +39,16 @@ library MarketsLib {
         mapping(address => mapping(uint256 => bool)) accountMembership;
     }
 
-    event MarketEntered(address fixedLender, address account, uint256 maturityDate);
-    event MarketExited(address fixedLender, address account, uint256 maturityDate);
+    event MarketEntered(
+        address fixedLender,
+        address account,
+        uint256 maturityDate
+    );
+    event MarketExited(
+        address fixedLender,
+        address account,
+        uint256 maturityDate
+    );
 
     /**
      * @dev Allows wallet to exit certain markets (fixedLenderDAI, fixedLenderETH, etc)
@@ -50,8 +58,15 @@ library MarketsLib {
      * @param who wallet that wants to exit a market/maturity
      * @param maturityDate poolID in which the wallet will stop using as collateral
      */
-    function exitMarket(Book storage book, address fixedLenderAddress, address who, uint256 maturityDate) external {
-        MarketsLib.Market storage marketToExit = book.markets[fixedLenderAddress];
+    function exitMarket(
+        Book storage book,
+        address fixedLenderAddress,
+        address who,
+        uint256 maturityDate
+    ) external {
+        MarketsLib.Market storage marketToExit = book.markets[
+            fixedLenderAddress
+        ];
 
         if (marketToExit.accountMembership[who][maturityDate] == false) {
             return;
@@ -60,10 +75,12 @@ library MarketsLib {
         delete marketToExit.accountMembership[who][maturityDate];
 
         // load into memory for faster iteration
-        IFixedLender[] memory userAssetList = book.accountAssets[who][maturityDate];
-        uint len = userAssetList.length;
-        uint assetIndex = len;
-        for (uint i = 0; i < len; i++) {
+        IFixedLender[] memory userAssetList = book.accountAssets[who][
+            maturityDate
+        ];
+        uint256 len = userAssetList.length;
+        uint256 assetIndex = len;
+        for (uint256 i = 0; i < len; i++) {
             if (userAssetList[i] == IFixedLender(fixedLenderAddress)) {
                 assetIndex = i;
                 break;
@@ -74,7 +91,9 @@ library MarketsLib {
         assert(assetIndex < len);
 
         // copy last item in list to location of item to be removed, reduce length by 1
-        IFixedLender[] storage storedList = book.accountAssets[who][maturityDate];
+        IFixedLender[] storage storedList = book.accountAssets[who][
+            maturityDate
+        ];
         storedList[assetIndex] = storedList[storedList.length - 1];
         storedList.pop();
 
@@ -103,7 +122,11 @@ library MarketsLib {
             revert GenericError(ErrorCode.MARKET_NOT_LISTED);
         }
 
-        if (!book.markets[fixedLenderAddress].accountMembership[borrower][maturityDate]) {
+        if (
+            !book.markets[fixedLenderAddress].accountMembership[borrower][
+                maturityDate
+            ]
+        ) {
             // only fixedLenders may call borrowAllowed if borrower not in market
             if (msg.sender != fixedLenderAddress) {
                 revert GenericError(ErrorCode.NOT_A_FIXED_LENDER_SENDER);
@@ -113,15 +136,18 @@ library MarketsLib {
             addToMarket(book, fixedLenderAddress, borrower, maturityDate);
 
             // it should be impossible to break the important invariant
-            assert(book.markets[fixedLenderAddress].accountMembership[borrower][maturityDate]);
+            assert(
+                book.markets[fixedLenderAddress].accountMembership[borrower][
+                    maturityDate
+                ]
+            );
         }
 
         uint256 borrowCap = book.borrowCaps[fixedLenderAddress];
         // Borrow cap of 0 corresponds to unlimited borrowing
         if (borrowCap != 0) {
-            uint256 totalBorrows = IFixedLender(fixedLenderAddress).getTotalMpBorrows(
-                maturityDate
-            );
+            uint256 totalBorrows = IFixedLender(fixedLenderAddress)
+                .getTotalMpBorrows(maturityDate);
             uint256 nextTotalBorrows = totalBorrows + borrowAmount;
             if (nextTotalBorrows >= borrowCap) {
                 revert GenericError(ErrorCode.MARKET_BORROW_CAP_REACHED);
@@ -147,18 +173,13 @@ library MarketsLib {
         address fixedLenderToSimulate,
         uint256 withdrawAmount,
         uint256 borrowAmount
-    )
-        external 
-        view
-        returns (
-            uint256,
-            uint256
-        )
-    {
+    ) external view returns (uint256, uint256) {
         AccountLiquidity memory vars; // Holds all our calculation results
 
         // For each asset the account is in
-        IFixedLender[] memory assets = book.accountAssets[account][maturityDate];
+        IFixedLender[] memory assets = book.accountAssets[account][
+            maturityDate
+        ];
         for (uint256 i = 0; i < assets.length; i++) {
             IFixedLender asset = assets[i];
             MarketsLib.Market storage market = book.markets[address(asset)];
@@ -168,28 +189,52 @@ library MarketsLib {
                 account,
                 maturityDate
             );
-            vars.collateralFactor = book.markets[address(asset)].collateralFactor;
+            vars.collateralFactor = book
+                .markets[address(asset)]
+                .collateralFactor;
 
             // Get the normalized price of the asset (18 decimals)
-            vars.oraclePrice = oracle.getAssetPrice(asset.underlyingTokenName());
+            vars.oraclePrice = oracle.getAssetPrice(
+                asset.underlyingTokenName()
+            );
 
             // We sum all the collateral prices
-            vars.sumCollateral += DecimalMath.getTokenAmountInUSD(vars.balance, vars.oraclePrice, market.decimals).mul_(vars.collateralFactor);
+            vars.sumCollateral += DecimalMath
+                .getTokenAmountInUSD(
+                    vars.balance,
+                    vars.oraclePrice,
+                    market.decimals
+                )
+                .mul_(vars.collateralFactor);
 
             // We sum all the debt
-            vars.sumDebt += DecimalMath.getTokenAmountInUSD(vars.borrowBalance, vars.oraclePrice, market.decimals);
+            vars.sumDebt += DecimalMath.getTokenAmountInUSD(
+                vars.borrowBalance,
+                vars.oraclePrice,
+                market.decimals
+            );
 
             // Simulate the effects of borrowing from/lending to a pool
             if (asset == IFixedLender(fixedLenderToSimulate)) {
                 // Calculate the effects of borrowing fixedLenders
                 if (borrowAmount != 0) {
-                    vars.sumDebt += DecimalMath.getTokenAmountInUSD(borrowAmount, vars.oraclePrice, market.decimals);
+                    vars.sumDebt += DecimalMath.getTokenAmountInUSD(
+                        borrowAmount,
+                        vars.oraclePrice,
+                        market.decimals
+                    );
                 }
 
                 // Calculate the effects of redeeming fixedLenders
                 // (having less collateral is the same as having more debt for this calculation)
                 if (withdrawAmount != 0) {
-                    vars.sumDebt += DecimalMath.getTokenAmountInUSD(withdrawAmount, vars.oraclePrice, market.decimals).mul_(vars.collateralFactor);
+                    vars.sumDebt += DecimalMath
+                        .getTokenAmountInUSD(
+                            withdrawAmount,
+                            vars.oraclePrice,
+                            market.decimals
+                        )
+                        .mul_(vars.collateralFactor);
                 }
             }
         }
@@ -209,10 +254,19 @@ library MarketsLib {
      * @param who address of the user that it will start participating in a market/maturity
      * @param maturityDate poolID in which it will start participating
      */
-    function addToMarket(Book storage book, address fixedLenderAddress, address who, uint256 maturityDate) public {
-        MarketsLib.Market storage marketToJoin = book.markets[fixedLenderAddress];
+    function addToMarket(
+        Book storage book,
+        address fixedLenderAddress,
+        address who,
+        uint256 maturityDate
+    ) public {
+        MarketsLib.Market storage marketToJoin = book.markets[
+            fixedLenderAddress
+        ];
         addToMaturity(marketToJoin, who, maturityDate);
-        book.accountAssets[who][maturityDate].push(IFixedLender(fixedLenderAddress));
+        book.accountAssets[who][maturityDate].push(
+            IFixedLender(fixedLenderAddress)
+        );
         emit MarketEntered(fixedLenderAddress, who, maturityDate);
     }
 
@@ -223,8 +277,12 @@ library MarketsLib {
      * @param borrower wallet that wants to enter a market
      * @param maturityDate poolID in which the wallet will be added to
      */
-    function addToMaturity(Market storage market, address borrower, uint256 maturityDate) internal {
-        if(!TSUtils.isPoolID(maturityDate)) { 
+    function addToMaturity(
+        Market storage market,
+        address borrower,
+        uint256 maturityDate
+    ) internal {
+        if (!TSUtils.isPoolID(maturityDate)) {
             revert GenericError(ErrorCode.INVALID_POOL_ID);
         }
 
@@ -238,5 +296,4 @@ library MarketsLib {
 
         market.accountMembership[borrower][maturityDate] = true;
     }
-
 }
